+++ logging/logrus/ctxlogrus/doc.go
@@ -1,0 +1,4 @@
+// ctxlogrus allows you to store or extract a logrus logger from the context.
+//
+// The tags on the logger are populated by http_ctxtags if they have already been set.
+package ctxlogrus
