--- tags/logrus/doc.go
@@ -1,4 +1,0 @@
-// ctx_logrus allows you to store or extract a logrus logger from the context.
-//
-// The tags on the logger are populated by http_ctxtags if they have already been set.
-package ctx_logrus
